Refactor gulpfile.js
