+++ util/options.js
@@ -1,0 +1,82 @@
+/*!
+ * Adorade (v1.0.0): options.js
+ * Copyright (c) 2018 Adorade (https://adorade.ro)
+ * Licensed under MIT
+ * ============================================================================
+ */
+
+const { dirs } = require('./config');
+
+const opts = {
+  styles: {
+    failAfterError: true,
+    reportOutputDir: 'logs/gulp',
+    reporters: [
+      { formatter: 'string', console: true, save: 'styles.txt' }
+    ],
+    syntax: 'scss'
+  },
+  sass: {
+    outputStyle: 'expanded',
+    precision: 6
+  },
+  autoprefixer: {
+    // for browsers options see .browserslistrc
+    cascade: false
+  },
+  csso: {
+    comments: false
+  },
+  eslint: {
+    // for more options see .eslintrc.json
+  },
+  babel: {
+    // for more options see .babelrc.js
+    comments: false
+  },
+  uglify: {
+    compress: {
+      evaluate: false
+    },
+    mangle: {
+      keep_fnames: true
+    }
+  },
+  images: {
+    gif: { interlaced: true },
+    jpeg: { progressive: true },
+    png: { optimizationLevel: 4 },
+    svg: { plugins: [{ removeViewBox: true }] },
+    webp: { // https://github.com/imagemin/imagemin-webp#options
+      preset: 'default',
+      quality: 60
+    }
+  },
+  pug: {
+    doctype: 'html',
+    pretty: true
+  },
+  html: {
+    collapseBooleanAttributes: true,
+    collapseInlineTagWhitespace: true,
+    collapseWhitespace: true,
+    minifyCSS: true,
+    minifyJS: true,
+    removeAttributeQuotes: true,
+    removeComments: true,
+    removeEmptyAttributes: true,
+    removeRedundantAttributes: true,
+    removeScriptTypeAttributes: true,
+    removeStyleLinkTypeAttributes: true
+  },
+  inline: {
+    rootpath: `${dirs.dest}/`
+  },
+  watch: {
+    delay: 2000
+  }
+};
+
+module.exports = {
+  opts
+};
